backend/requirements.txt

django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pandas
reportlab
sqlite3
